refactor : BettingResult 삭제
